requirements.txt
absl-py==2.2.2
accelerate==1.6.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
albumentations==0.5.2
antlr4-python3-runtime==4.8
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
av==13.0.0
braceexpand==0.1.7
cloudpickle==3.1.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
cython==3.0.11
decord==0.6.0
diffdist==0.1
diffusers==0.34.0
easydict==1.9
einops==0.4.1
executing==2.2.0
facexlib==0.3.0
filterpy==1.4.5
flash-attn==2.7.4.post1
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
ftfy==6.1.1
future==1.0.0
gfpgan==1.3.8
grpcio==1.71.0
h5py==3.7.0
huggingface-hub==0.30.2
humanfriendly==10.0
hydra-core==1.1.0
icecream==2.1.2
imageio==2.37.0
imageio-ffmpeg==0.4.9
imgaug==0.4.0
importlib-metadata==8.6.1
inflect==5.6.0
joblib==1.4.2
kiwisolver==1.4.8
kornia==0.8.0
kornia-rs==0.1.8
lazy-loader==0.4
lightning-utilities==0.14.3
llvmlite==0.44.0
lmdb==1.6.2
loguru==0.7.3
markdown==3.8
markdown-it-py==3.0.0
matplotlib==3.10.1
mdurl==0.1.2
multidict==6.4.3
mmcv==1.7.0
mmengine==0.10.7
numba==0.61.2
numpy==1.24.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.1.2
onnx==1.17.0
onnxruntime-gpu==1.21.1
open-clip-torch==2.30.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
peft==0.15.0
platformdirs==4.3.7
plyfile==1.1
propcache==0.3.1
protobuf==5.29.3
psutil==7.0.0
py-cpuinfo==9.0.0
py360convert==1.0.3
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2025.2
qwen-vl-utils==0.0.8
regex==2022.6.2
rich==14.0.0
safetensors==0.5.3
sageattention==1.0.6
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
segment-anything==1.0
sentencepiece==0.2.0
setuptools==59.5.0
shapely==2.0.7
six==1.17.0
submitit==1.4.2
sympy==1.13.1
tabulate==0.9.0
tb-nightly==2.20.0a20250421 -i https://mirrors.aliyun.com/pypi/simple
tensorboard-data-server==0.7.2
termcolor==3.0.1
threadpoolctl==3.6.0
tifffile==2025.3.30
timm==1.0.13
tokenizers==0.21.1
tomli==2.2.1
torchmetrics==1.7.1
tqdm==4.67.1
transformers==4.51.0
triton==3.2.0
tzdata==2025.2
ultralytics==8.3.74
ultralytics-thop==2.0.14
wcwidth==0.2.13
webdataset==0.2.100
werkzeug==3.1.3
wldhx-yadisk-direct==0.0.6
xformers==0.0.28.post2
yapf==0.43.0
yarl==1.20.0
zipp==3.21.0
open3d>=0.18.0
trimesh>=4.6.1
cmake
git+https://github.com/EasternJournalist/utils3d.git@3fab839f0be9931dac7c8488eb0e1600c236e183